requirements.txt
aiohttp==3.7.4.post0
astroid==2.11.3
async-timeout==3.0.1
attrs==21.4.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
colorama==0.4.4
cycler==0.11.0
dill==0.3.4
discord-py-interactions==4.1.0
discord.py==1.7.3
fonttools==4.37.1
idna==3.3
isort==5.10.1
kiwisolver==1.4.4
lazy-object-proxy==1.7.1
matplotlib==3.5.3
mccabe==0.7.0
pylint==2.13.7
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.27.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
wrapt==1.14.0
yarl==1.7.2
